SqrMelon/requirements.txt
setuptools
PyOpenGL==3.1.0
pyprof2calltree==1.4.3
pyOSC==0.3.5b5294
Send2Trash==1.5.0
pyglet==1.2.4
pygame==1.9.3
